CodeQL warning solved

+++ units/vector_unit_2d.cpp
@@ -389,7 +389,7 @@ namespace scifir
 
 scifir::vector_unit_2d operator *(const scifir::scalar_unit& x,const scifir::vector_unit_2d& y)
 {
-	long double new_value = x.get_value() * y.get_value();
+	long double new_value = (long double)x.get_value() * (long double)y.get_value();
 	vector<scifir::dimension> new_dimensions = multiply_dimensions(x.get_dimensions(), y.get_dimensions(),new_value);
 	return scifir::vector_unit_2d(float(new_value), new_dimensions, y.theta);
 }
